feat: consolidate views
